Assignment-1-Sketch-to-image/requirements.txt
requests
Pillow
flask-cors